Refactor and enhance codebase for improved functionality and security
- Simplified legacy password retrieval in KeychainStore.
- Introduced a default implementation for KeychainStore.
- Updated sync host key policy to enforce known hosts for secure SFTP sync.
- Improved error handling in SFTP commands and added validation for remote leaf names.
- Enhanced local directory removal to prevent unsafe operations.
- Added tests for path validation and remote directory copy operations.
- Updated Tauri configuration to tighten Content Security Policy (CSP).

+++ core/crates/zeroterm-sync/src/repo.rs
@@ -142,7 +142,7 @@ fn year_month(unix_ms: i64) -> (u32, u32) {
     // Lightweight days-since-epoch → (Y, M) conversion. Good through 2400
     // and gives stable bucket names independent of the local timezone.
     // Algorithm: Howard Hinnant's `civil_from_days` (variant Rata Die).
-    let days = (unix_ms.max(0) / 86_400_000) as i64 + 719_468; // 1970-01-01 → 719468
+    let days = unix_ms.max(0) / 86_400_000 + 719_468; // 1970-01-01 → 719468
     let era = if days >= 0 { days } else { days - 146_096 } / 146_097;
     let doe = (days - era * 146_097) as u32; // [0, 146097)
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365;
